+++ app/Http/Controllers/UsersController.php
@@ -2,9 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\RegistrationRequest;
+use App\Http\Requests\UserRequest;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Inertia\Inertia;
 
 class UsersController extends Controller
 {
@@ -18,7 +19,10 @@ class UsersController extends Controller
      */
     public function show(User $user)
     {
-        return view('users.show', ['user' => $user->with(['posts', 'post_collections'])->get()]);
+        return Inertia::render(
+            'Users/Show',
+            ['user' => $user->with(['posts', 'postCollections'])->get()]
+        );
     }
 
     /**
@@ -29,25 +33,27 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
-        return view('users.edit', ['user' => $user]);
+        $this->authorize('update', $user);
+        return Inertia::render('Users/Edit', ['user' => $user]);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \App\Http\Requests\RegistrationRequest $request
+     * @param  \App\Http\Requests\UserRequest $request
      * @param  \App\Models\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function update(RegistrationRequest $request, User $user)
+    public function update(UserRequest $request, User $user)
     {
+        $this->authorize('update', $user);
         if ($user->update($request->validated())) {
             return redirect()
-                ->route('users.show', ['user' => $user])
+                ->route('users.show', ['user' => $user], 303)
                 ->with('success', 'Profile updated successfully!');
         } else {
             return redirect()
-                ->route('users.edit', ['user' => $user])
+                ->route('users.edit', ['user' => $user], 303)
                 ->with('error', 'Profile update failed!');
         }
     }
@@ -60,8 +66,12 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
         Auth::logout();
         Auth::logoutOtherDevices($user->password);
         $user->delete();
+        return redirect()
+            ->route('index', [], 303)
+            ->with('success', 'Account deleted successfully!');
     }
 }
